[#100] 관광 정보 버그 수정 및 관광지 상세보기 화면 이동

+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/model/Place.kt
@@ -5,5 +5,5 @@ data class Place (
     val disability: List<String>,
     val image: String,
     val name: String,
-    val placeId: Int,
+    val placeId: Long
 )
